Add support for MojoLoader-based frameworks
Loads mojoloader.js if the config specifies that mojoloader is being used
(via 'useMojoLoader' in eris_config.json).

The config variable 'localEnyoRoot' is intended to refer specifically to the
directory where the Enyo framework is kept, and implicitly requires that the
directory 'enyo' be contained in that directory.  This change makes this state
of affairs explicit in the code, which frees the mapping for /usr/palm/frameworks
to refer to frameworks more generally.

Adds support for 'localFrameworksRoot' in eris_config.json; MojoLoader-based
frameworks are often checked out in a different directory than Enyo on
development machines.  This is also the location of mojoloader.js.